+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 - Implementation code moved to [Eozilla](https://github.com/eo-tools/eozilla) (#39):
   - Package `s2gos-client` now uses the Eozilla package `cuiman`.
+  - Package `s2gos-server` now uses the Eozilla package `wraptile`.
   - Package `s2gos-app-ex` now uses the Eozilla package `procodile`.
 
 ## Changes in version 0.0.5